Break import cycle by extracting MFA JSON types to lib/client/mfatypes
