+++ lib/client/weblogin.go
@@ -46,6 +46,7 @@ import (
 	"github.com/gravitational/teleport/lib/auth/authclient"
 	wancli "github.com/gravitational/teleport/lib/auth/webauthncli"
 	wantypes "github.com/gravitational/teleport/lib/auth/webauthntypes"
+	"github.com/gravitational/teleport/lib/client/mfatypes"
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/httplib"
 	websession "github.com/gravitational/teleport/lib/web/session"
@@ -119,89 +120,16 @@ type MFAChallengeRequest struct {
 }
 
 // MFAChallengeResponse holds the response to a MFA challenge.
-type MFAChallengeResponse struct {
-	// TOTPCode is a code for a otp device.
-	TOTPCode string `json:"totp_code,omitempty"`
-	// WebauthnResponse is a response from a webauthn device.
-	WebauthnResponse *wantypes.CredentialAssertionResponse `json:"webauthn_response,omitempty"`
-	// SSOResponse is a response from an SSO MFA flow.
-	SSOResponse *SSOResponse `json:"sso_response"`
-	// TODO(Joerger): DELETE IN v20.0.0, WebauthnResponse used instead.
-	WebauthnAssertionResponse *wantypes.CredentialAssertionResponse `json:"webauthnAssertionResponse"`
-	// BrowserMFAResponse is a response the browser completing an MFA challenge
-	// as part of the Browser MFA flow.
-	BrowserMFAResponse *BrowserMFAResponse `json:"browser_response"`
-}
+type MFAChallengeResponse = mfatypes.MFAChallengeResponse
 
 // SSOResponse is a json compatible [proto.SSOResponse].
-type SSOResponse struct {
-	RequestID string `json:"requestId,omitempty"`
-	Token     string `json:"token,omitempty"`
-}
+type SSOResponse = mfatypes.SSOResponse
 
 // BrowserMFAResponse is a json compatible [proto.BrowserMFAResponse].
-type BrowserMFAResponse struct {
-	RequestID        string                                `json:"requestId,omitempty"`
-	WebauthnResponse *wantypes.CredentialAssertionResponse `json:"webauthnResponse,omitempty"`
-}
-
-// GetOptionalMFAResponseProtoReq converts response to a type proto.MFAAuthenticateResponse,
-// if there were any responses set. Otherwise returns nil.
-func (r *MFAChallengeResponse) GetOptionalMFAResponseProtoReq() (*proto.MFAAuthenticateResponse, error) {
-	if r == nil {
-		return nil, nil
-	}
-
-	var availableResponses int
-	if r.TOTPCode != "" {
-		availableResponses++
-	}
-	if r.WebauthnResponse != nil {
-		availableResponses++
-	}
-	if r.SSOResponse != nil {
-		availableResponses++
-	}
-
-	if availableResponses > 1 {
-		return nil, trace.BadParameter("only one MFA response field can be set")
-	}
-
-	switch {
-	case r.WebauthnResponse != nil:
-		return &proto.MFAAuthenticateResponse{Response: &proto.MFAAuthenticateResponse_Webauthn{
-			Webauthn: wantypes.CredentialAssertionResponseToProto(r.WebauthnResponse),
-		}}, nil
-	case r.SSOResponse != nil:
-		return &proto.MFAAuthenticateResponse{Response: &proto.MFAAuthenticateResponse_SSO{
-			SSO: &proto.SSOResponse{
-				RequestId: r.SSOResponse.RequestID,
-				Token:     r.SSOResponse.Token,
-			},
-		}}, nil
-	case r.TOTPCode != "":
-		return &proto.MFAAuthenticateResponse{Response: &proto.MFAAuthenticateResponse_TOTP{
-			TOTP: &proto.TOTPResponse{Code: r.TOTPCode},
-		}}, nil
-	case r.WebauthnAssertionResponse != nil:
-		return &proto.MFAAuthenticateResponse{Response: &proto.MFAAuthenticateResponse_Webauthn{
-			Webauthn: wantypes.CredentialAssertionResponseToProto(r.WebauthnAssertionResponse),
-		}}, nil
-	}
-
-	return nil, nil
-}
+type BrowserMFAResponse = mfatypes.BrowserMFAResponse
 
 // ParseMFAChallengeResponse parses [MFAChallengeResponse] from JSON and returns it as a [proto.MFAAuthenticateResponse].
-func ParseMFAChallengeResponse(mfaResponseJSON []byte) (*proto.MFAAuthenticateResponse, error) {
-	var resp MFAChallengeResponse
-	if err := json.Unmarshal(mfaResponseJSON, &resp); err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	protoResp, err := resp.GetOptionalMFAResponseProtoReq()
-	return protoResp, trace.Wrap(err)
-}
+var ParseMFAChallengeResponse = mfatypes.ParseMFAChallengeResponse
 
 // HeadlessLoginReq is a headless login request for /webapi/headless/login.
 type HeadlessLoginReq struct {
@@ -416,46 +344,22 @@ type SSHLoginHeadless struct {
 
 // MFAAuthenticateChallenge is an MFA authentication challenge sent on user
 // login / authentication ceremonies.
-type MFAAuthenticateChallenge struct {
-	// WebauthnChallenge contains a WebAuthn credential assertion used for
-	// login/authentication ceremonies.
-	WebauthnChallenge *wantypes.CredentialAssertion `json:"webauthn_challenge"`
-	// TOTPChallenge specifies whether TOTP is supported for this user.
-	TOTPChallenge bool `json:"totp_challenge"`
-	// SSOChallenge is an SSO MFA challenge.
-	SSOChallenge *SSOChallenge `json:"sso_challenge"`
-	// BrowserMFAChallenge is a Browser MFA challenge.
-	BrowserMFAChallenge *BrowserMFAChallenge `json:"browser_challenge"`
-}
+type MFAAuthenticateChallenge = mfatypes.MFAAuthenticateChallenge
 
 // SSOChallenge is a json compatible [proto.SSOChallenge].
-type SSOChallenge struct {
-	RequestID   string        `json:"requestId,omitempty"`
-	RedirectURL string        `json:"redirectUrl,omitempty"`
-	Device      *SSOMFADevice `json:"device"`
-	// ChannelID is used by the front end to differentiate multiple ongoing SSO
-	// MFA requests so they don't interfere with each other.
-	ChannelID string `json:"channelId"`
-}
+type SSOChallenge = mfatypes.SSOChallenge
 
 // SSOMFADevice is a json compatible [proto.SSOMFADevice].
-type SSOMFADevice struct {
-	ConnectorID   string `json:"connectorId,omitempty"`
-	ConnectorType string `json:"connectorType,omitempty"`
-	DisplayName   string `json:"displayName,omitempty"`
-}
+type SSOMFADevice = mfatypes.SSOMFADevice
 
-func SSOChallengeFromProto(ssoChal *proto.SSOChallenge) *SSOChallenge {
-	return &SSOChallenge{
-		RequestID:   ssoChal.RequestId,
-		RedirectURL: ssoChal.RedirectUrl,
-		Device: &SSOMFADevice{
-			ConnectorID:   ssoChal.Device.ConnectorId,
-			ConnectorType: ssoChal.Device.ConnectorType,
-			DisplayName:   ssoChal.Device.DisplayName,
-		},
-	}
-}
+// BrowserMFAChallenge is a json compatible [proto.BrowserMFAChallenge].
+type BrowserMFAChallenge = mfatypes.BrowserMFAChallenge
+
+var (
+	SSOChallengeFromProto     = mfatypes.SSOChallengeFromProto
+	BrowserChallengeToProto   = mfatypes.BrowserChallengeToProto
+	BrowserChallengeFromProto = mfatypes.BrowserChallengeFromProto
+)
 
 // MFARegisterChallenge is an MFA register challenge sent on new MFA register.
 type MFARegisterChallenge struct {
@@ -468,25 +372,6 @@ type MFARegisterChallenge struct {
 // TOTPRegisterChallenge contains a TOTP challenge.
 type TOTPRegisterChallenge struct {
 	QRCode []byte `json:"qrCode"`
-}
-
-// BrowserMFAChallenge is a json compatible [proto.BrowserMFAChallenge].
-type BrowserMFAChallenge struct {
-	RequestID string `json:"requestId,omitempty"`
-}
-
-// BrowserChallengeToProto converts an BrowserChallenge to proto format.
-func BrowserChallengeToProto(browserChal *BrowserMFAChallenge) *proto.BrowserMFAChallenge {
-	return &proto.BrowserMFAChallenge{
-		RequestId: browserChal.RequestID,
-	}
-}
-
-// BrowserChallengeFromProto converts a BrowserChallenge to json compatible format
-func BrowserChallengeFromProto(browserChal *proto.BrowserMFAChallenge) *BrowserMFAChallenge {
-	return &BrowserMFAChallenge{
-		RequestID: browserChal.RequestId,
-	}
 }
 
 // initClient creates a new client to the HTTPS web proxy.
